minor updates
attempted to add a GUI for agent health, unable to accomplish. Would be
difficult since we are removing agents from list as they die or get to
waypoint
Currently, health for agents is set to 1 to test agent death.  Upon
reset of the game, their health is still 0 so they are not used. Need to
decide whether they still spawn again if killed, if with more health, or
is that the end of the game?